file manager/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
chromadb
python-docx
docx2txt
pypdf2
tqdm
watchdog
